requirements.txt
numpy
Pillow
Flask
RPi.GPIO
spidev
pytest